fix: set directory as deleted

+++ src/server/udp/udp_commands.c
@@ -65,8 +65,14 @@ void udp_delete_user(char * arg1, char * arg2, udp_resp * resp) {
         return;
     }
 
+    if (user->deleted) {
+        resp->code = SERVER_ERROR;
+        return;
+    }
+
     if (delete_user_dir(username, strlen(username)) == 0) {
         resp->code = OK;
+        user->deleted = true;
         return;
     }
